+++ CHANGELOG.md
@@ -24,3 +24,7 @@
 * Rewrite README with a better english
 * Added == hashcode and toString method to the PagedNotifier
 * Added documentation on single classes
+
+## 1.0.6
+
+* Format code